feat: Revamped the project structure and improved documentation
