+++ .env.example
@@ -1,0 +1,8 @@
+# Environment Configuration
+# Copy this file to .env.local for local development
+# Copy this file to .env.production for production builds
+
+# Base path for the application
+# For local development, use "/"
+# For GitHub Pages, use "/repository-name/"
+VITE_APP_BASE_PATH=/
