+++ README.md
@@ -1,6 +1,6 @@
 ## ℹ️ Barcode Scanner with ZBar WASM
 
-This project is a barcode scanner application built with React and ZBar WASM. It uses the device's camera to scan barcode and display the decoded information.
+This project is a barcode scanner application built with React and ZBar WASM. It uses the device's camera to scan barcodes and display the decoded information.
 
 ## 👨🏻‍💻 Live App
 
@@ -8,13 +8,59 @@ You can try the live version on GitHub pages here: https://sumitsahoo.github.io/
 
 ## 📖 Features
 
-1. PWA app can be installed on any device (on supported browsers)
-2. Real-time barcode scanning
-3. Start and stop scanning functionality
-4. Switch between front and back camera in phones and tablets
-5. Toggle flash for night time scanning
-6. Display of scanning result in a dialog box with copy option
-7. Haptic feedback (Android) on successful scan
+1. **PWA Support** - Can be installed on any device (on supported browsers)
+2. **Real-time Scanning** - Fast and efficient barcode detection
+3. **Camera Controls** - Start, stop, and switch between front/back cameras
+4. **Flash Toggle** - Use torch for scanning in low-light conditions
+5. **Result Dialog** - Display scan results with copy-to-clipboard functionality
+6. **Haptic Feedback** - Vibration feedback on successful scan (mobile devices)
+7. **Responsive Design** - Works seamlessly on mobile and desktop
+8. **Error Handling** - Graceful error boundaries for better UX
+
+## 🏗️ Architecture
+
+### Project Structure
+
+```
+src/
+├── components/
+│   ├── BarcodeScanner/      # Main scanner component
+│   │   ├── BarcodeScanner.jsx
+│   │   ├── ScannerControls.jsx
+│   │   ├── ResultDialog.jsx
+│   │   └── index.js
+│   ├── ErrorBoundary.jsx    # Error boundary wrapper
+│   └── icons/               # SVG icon components
+├── hooks/
+│   └── useBarcodeScanner.js # Scanner logic custom hook
+├── utils/
+│   ├── barcodeHelpers.js    # Barcode scanning utilities
+│   ├── themeColors.js       # Theme color utilities
+│   └── frameBuster.js       # Clickjacking prevention
+├── constants/
+│   ├── scanner.js           # Scanner configuration
+│   └── camera.js            # Camera settings
+├── config/
+│   └── env.js               # Environment validation
+└── main.jsx                 # Application entry point
+```
+
+### Key Technologies
+
+- **React 19** - Latest React with performance optimizations
+- **Vite** - Fast build tool and dev server
+- **ZBar WASM** - WebAssembly barcode detection library
+- **Tailwind CSS + DaisyUI** - Utility-first styling framework
+- **Biome** - Fast formatter and linter
+- **PWA** - Progressive Web App capabilities
+
+### Design Patterns
+
+- **Custom Hooks** - Business logic separated into reusable hooks
+- **Component Composition** - Small, focused components
+- **Error Boundaries** - Graceful error handling
+- **Memoization** - Optimized performance with React.memo, useMemo, useCallback
+- **Lazy Loading** - ZBar library loaded on-demand
 
 ## 📷 Screenshots
 
@@ -26,11 +72,9 @@ You can try the live version on GitHub pages here: https://sumitsahoo.github.io/
 
 ![iPad](public/screenshots/screenshot-2732x2048.png "App running on iPad Pro")
 
-
 ## 🧐 Why this is not a library?
 
 In many scenarios, the user interface (UI) requirements for scanner implementations vary significantly. This diversity in UI needs makes it impractical to encapsulate such functionality within a library, as it would severely limit the ability to customize the UI to meet specific requirements.
-
 
 ## 📦 Local Installation
 
@@ -38,6 +82,7 @@ Clone the repository:
 
 ```bash
 git clone https://github.com/sumitsahoo/barcode-scanner.git
+cd barcode-scanner
 ```
 
 Install the dependencies:
@@ -54,7 +99,16 @@ Start the application:
 npm run dev
 ```
 
-The application will open in your default web browser. Click the "Camera" button icon to begin scanning a barcode. The decoded information will be displayed in a dialog box.
+The application will open in your default web browser. Click the camera button icon to begin scanning a barcode. The decoded information will be displayed in a dialog box.
+
+### Available Scripts
+
+- `npm run dev` - Start development server
+- `npm run build` - Build for production
+- `npm run preview` - Preview production build
+- `npm run format` - Format code with Biome
+- `npm run lint` - Lint code with Biome
+- `npm run deploy` - Deploy to GitHub Pages
 
 ## 🚢 Docker
 
@@ -85,7 +139,9 @@ barcode-scanner
 
 ## 💼 Contributing
 
-Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.
+We welcome contributions! Please see [CONTRIBUTING.md](CONTRIBUTING.md) for detailed guidelines.
+
+For major changes, please open an issue first to discuss what you would like to change.
 
 ## 🙏🏻 Attributions
