+++ package.json
@@ -2,7 +2,7 @@
   "name": "barcode-scanner",
   "author": "Sumit Sahoo",
   "private": true,
-  "version": "1.0.5",
+  "version": "1.0.6",
   "type": "module",
   "homepage": "https://sumitsahoo.github.io/barcode-scanner",
   "scripts": {
@@ -12,7 +12,8 @@
     "predeploy": "npm run build",
     "deploy": "gh-pages -d dist",
     "format": "npx biome format ./ --write",
-    "lint": "npx biome lint ./ --apply",
+    "lint": "npx biome lint ./",
+    "lint:fix": "npx biome lint ./ --write",
     "generate-pwa-assets": "pwa-assets-generator --preset minimal public/images/ic-logo.svg"
   },
   "dependencies": {
@@ -27,8 +28,8 @@
     "@biomejs/biome": "2.2.5",
     "@tailwindcss/postcss": "^4.1.14",
     "@tailwindcss/vite": "^4.1.14",
-    "@types/react": "^19.2.1",
-    "@types/react-dom": "^19.2.0",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.1",
     "@vite-pwa/assets-generator": "^1.0.1",
     "@vitejs/plugin-react": "^5.0.4",
     "daisyui": "^5.1.27",
