+++ src/components/BarcodeScanner/index.js
@@ -1,0 +1,4 @@
+/**
+ * Barrel export for BarcodeScanner component
+ */
+export { default } from "./BarcodeScanner";
