+++ tailwind.config.js
@@ -1,8 +1,8 @@
 module.exports = {
-  content: [
-    "./index.html",
-    "./src/**/*.{vue,js,ts,jsx,tsx}",
-    "node_modules/daisyui/dist/**/*.js",
-  ],
-  plugins: [require("daisyui")],
+	content: [
+		"./index.html",
+		"./src/**/*.{vue,js,ts,jsx,tsx}",
+		"node_modules/daisyui/dist/**/*.js",
+	],
+	plugins: [require("daisyui")],
 };
